+++ spec/commands/reserve_purchase_spec.rb
@@ -16,11 +16,11 @@
 
 require "rails_helper"
 
-RSpec.describe PurchaseMembership do
+RSpec.describe ReservePurchase do
   let(:membership_number) { nil }
   let(:membership) { create(:membership, :adult) }
   let(:user) { create(:user) }
-  let(:command) { PurchaseMembership.new(membership, customer: user, membership_number: membership_number) }
+  let(:command) { ReservePurchase.new(membership, customer: user, membership_number: membership_number) }
 
   context "when successful" do
     it "returns true" do